Replaced Lines of Code with Late Night Runaway.
Lines of Code, it turns out, gets painful to listen to after a while.
This wouldn't be a problem if it was restarting a lot (e.g. due to
having traditional levels), but that's not the case in Hexoshi, so
this is a serious problem. It's a shame; I liked Lines of Code, but
Late Night Runaway is certainly a worthy replacement.