Added a message when you first activate a warp pad.
This message makes it clearer what the warp pad does. The fact that
it saves progress was simple enough, but the fact that you can warp
between them was not so obvious and will be important later in the
game. It's also the only way to see the map before you acquire the
handheld map powerup, so that's a good thing to know.