Added back in Lines of Code for use as credits music.
Lines of Code has *terrible* loudness war syndrome. Still, it sounds
really great for a short period of time, so I think it's a good
choice for the credits music. I've added it back in for use there.
The original "fade-in/fade-out" version is used rather than the
custom looping version I had.