Added some useless artifacts to scatter around the world.
After much deliberation, I've decided to go through with the idea.
I won't be putting tons of them in the game, but I will put
useless artifacts in passages that are designed to get you lost, so
that they are not *entirely* meaningless.

Regarding today's work, the only part that I timed was the creation
of the new room, which took 10 minutes. I didn't time the rest of it
partly because it had very little to do with the goals of Milestone
1 and partly because I just forgot.