Changed the E-tank in room 5 so that you can get it with the spike ball.
The original plan for this room was to require the Sloth Ball, but
seeing the technical challenges with that, I will probably not be
doing that. So instead, this E-tank will be intended to be made
accessible by a powerup later in the game which causes spikes to
grow out of you when you jump, and you can get it early by doing a
well-timed jump with theh spike ball.