Implemented the Mantanoid enemy.
My God, I'm so excited! It actually works really well! The Mantanoid
charges at you like a maniac, and you are basically forced to run
for your life.

I had to fix up the sprite sheet, and I might want to do some more
tweaks as well. Other than that, it's just adding sound effects, and
putting the Mantanoids in a couple areas of Iridia.