More modifications to machine learning.
This was a loooooooong and complicated process, but basically, I did
a whole bunch of testing and tried to optimize it so that it would
learn better. Lots of things I tested but ended up throwing away,
etc. I think I've come up with something that works better at the
end of it, though.

Also tweaked room 9 to prevent the Mantanoid from getting stuck in
the corner it comes out of (the A.I. just can't handle that sort
of complexity unfortunately).