Added a rudimentary curiosity system.
Still needs work. These "IDs" aren't going to work. We have to store
more info about how well each action went so that the AI can decide
on a hierarchy of "best action" to "worst action" so that the AI can
start to prioritize actions that get it much closer than known actions,
and actions that have very little success can lead to a higher rate of
curiosity.