Some notes about Git repo in readme, plus warn rather than error on mapgen
The map generation writes to DATA, so if Hexoshi is not in a
user-writable location, that could result in a PermissionError. This
change takes that into an account by catching PermissionError and
spitting out a warning rather than raising an exception. This may
help if for some reason a distro forgets to package the map files.