Changed resolution from 400x224 to 400x240, plus some fixes.
The resolution change tool all day, but I think it's worth it
because it makes the vertical resolution less irregular, even if it
means having a 5:3 (15:9) aspect ratio instead of 16:9 (which it
actually wasn't anyhow; 400x224 is 25:14, or about 16.07:9; at
1080p resolution the effective display resolution would be 1920x1075).
In particular, it's worth it because it means the vertical resolution
fits neatly into 720p resolution (scaling at exactly 3x for a display
resolution of 1200x720), as well as 480p widescreen (scaling at exactly
2x for a display resolution of 800x480). It's also close enough to
16:9 that the image can be stretched while leaving the aspect ratio
change relatively unnoticeable. We'll be adding an option for this
later on.