Started work on a new starting room.
This new starting room will be laid out completely differently and
designed specifically as a tutorial of how the game works. Going to
the right takes you to a dead-end (which will be accessible only
after getting the Hedgehog Hormone). Going to the left will be designed
to take you to the door on the left, another dead-end. The player will
be tasked with finding the Hex Orb to proceed.