Adjusted the camera to make it scroll vertically better.
A complaint I saw was that the camera system doesn't give you much
vertical view, which has a slightly complicated cause. I experimented
with a few methods, but ultimately settled on retaining the system
roughly as it was, just with number adjustments and removal of the
Super Mario World style vertical locking of the camera while in the
air. I also changed the method used for camera guides: instead of
checking for collision with the player, it checks for collision with
a 1 pixel rectangle at the player's position. This prevents weird
variance with things that change the player's size.