Removed the unused infrastructure for multiplayer, replaced mode setting.
This was a lot of effort, but makes code a lot cleaner. The multiplayer
controls infrastructure was a leftover from an idea in ReTux for
simultaneous multiplayer (like New Super Mario Bros Wii). Multiplayer
isn't off the table for Hexoshi entirely, but if it happens, it'll
work in a manner more similar to Sonic 2's usage of controller 2 for
Tails, so it doesn't make sense for player 2's controls to be a copy
of player 1's controls.

While I made this change, I simultaneously removed the "mode" control
that was previously in use, replacing it with a more Mega Man Zero-like
"primary" and "secondary" control method. This I think will be a far
better control system in the end since cycling through various weapons
can be rather tedious. The only downside to this approach is a limited
number of available weapons, but if this becomes a problem, we can always
implement a weapons selection menu similar to Mega Man Zero. That said,
I don't anticipate implementing large numbers of different weapons.